fix: broken student search for "Pickup Poster" (/printing/posterPickup)
It looks like from the upgrade from PHP 5 to PHP 7 or upgrading Laravel
might have caused this issue.

`Input::get(key)` will return NULL if the input is an empty string.

The call to `->orWhere('foo', 'LIKE', NULL)` will generate
`OR foo IS NOT NULL`.

So before if any of the inputs are blank, the student search query would
become `... OR column IS NOT NULL OR ...` which would be true for all of
the rows.

This fix restores the original intent of the code which instead becomes
for example `... OR FirstName LIKE '' ...` so that only the search
fields specified will matter.